+++ src/main/frontend/src/components/DatePickerSanjy.tsx
@@ -1,0 +1,5 @@
+import { DateInput, type DateInputProps } from '@mantine/dates';
+
+export function DatePickerSanjy(props: Readonly<DateInputProps>) {
+  return <DateInput clearable {...props} />;
+}