+++ src/main/frontend/src/hooks/useGetAppTitleRedirectPath.ts
@@ -1,0 +1,6 @@
+import { useAppRuntimeConfig } from '../contexts/AppRuntimeConfigContext.tsx';
+
+export function useGetAppTitleRedirectPath(): string {
+  const { runtimeConfiguration } = useAppRuntimeConfig();
+  return runtimeConfiguration.appTitleRedirectPath.value;
+}