+++ src/main/frontend/src/hooks/useGetLogoutUrl.ts
@@ -1,0 +1,6 @@
+import { useAppRuntimeConfig } from '../contexts/AppRuntimeConfigContext.tsx';
+
+export function useGetLogoutUrl(): string | null {
+  const { runtimeConfiguration } = useAppRuntimeConfig();
+  return runtimeConfiguration.logoutUrl.value;
+}