+++ src/main/java/br/com/gorillaroxo/sanjy/client/web/controller/dto/response/DietPlanControllerResponseDto.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDate;
 import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
-import java.util.Set;
 import lombok.Builder;
 
 @Builder(toBuilder = true)
@@ -97,7 +97,7 @@ public record DietPlanControllerResponseDto(
                 description = "List of meal types associated with this diet plan",
                 requiredMode = Schema.RequiredMode.REQUIRED)
         @JsonPropertyDescription("List of meal types associated with this diet plan")
-        Set<MealTypeControllerResponseDto> mealTypes,
+        List<MealTypeControllerResponseDto> mealTypes,
 
         @Schema(
                 description = "Indicates whether this diet plan is currently active",
@@ -114,6 +114,6 @@ public record DietPlanControllerResponseDto(
         MetadataControllerResponseDto metadata) {
 
     public DietPlanControllerResponseDto {
-        mealTypes = Objects.requireNonNullElse(mealTypes, Collections.emptySet());
+        mealTypes = Objects.requireNonNullElse(mealTypes, Collections.emptyList());
     }
 }