+++ src/main/java/br/com/gorillaroxo/sanjy/client/web/controller/dto/response/FrontendRuntimeConfigurationControllerResponseDto.java
@@ -1,0 +1,10 @@
+package br.com.gorillaroxo.sanjy.client.web.controller.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record FrontendRuntimeConfigurationControllerResponseDto(
+        RuntimeConfigEntryDto logoutUrl, RuntimeConfigEntryDto appTitleRedirectPath) {
+
+    public record RuntimeConfigEntryDto(String env, String value) {}
+}