+++ src/test/java/br/com/gorillaroxo/sanjy/client/web/test/IntegrationTestController.java
@@ -10,6 +10,7 @@ import br.com.gorillaroxo.sanjy.client.web.test.mockwebserver.MockWebServerDispa
 import br.com.gorillaroxo.sanjy.client.web.test.mockwebserver.MockWebServerManager;
 import br.com.gorillaroxo.sanjy.client.web.util.JsonUtil;
 import lombok.extern.slf4j.Slf4j;
+import org.jspecify.annotations.Nullable;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestInstance;
@@ -17,7 +18,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.webtestclient.autoconfigure.AutoConfigureWebTestClient;
 import org.springframework.context.annotation.Import;
-import org.springframework.lang.Nullable;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;